+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # [0.3.0] - Unreleased
 - Made feature a property of the like button.
 - Changed feature list to be source-based.
+- Added easeTo and flyTo methods to gl-map.
 
 # [0.2.1] - 2018-06-01
 - Ensure that gl-form-page is loaded by dependent modules.